noam - added changes to fusionSlam including calculations and updating landmarks. tamar- working on json

+++ src/main/java/bgu/spl/mics/application/objects/CloudPoint.java
@@ -10,7 +10,7 @@ public class CloudPoint {
     private final double x;
     private final double y;
 
-    public CloudPoint(int x, int y) {
+    public CloudPoint(double x, double y) {
         this.x = x;
         this.y = y;
     }
